requirements.txt
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.1.0
Deprecated==1.2.13
exceptiongroup==1.1.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.2.1
pytest-cov==4.0.0
requests==2.28.2
smmap==5.0.0
tomli==2.0.1
urllib3==1.26.14
wrapt==1.14.1
